relay_quotas

Enum DataCategory

#[repr(i8)]
pub enum DataCategory {

    Default = 0,
    Error = 1,
    Transaction = 2,
    Security = 3,
    Attachment = 4,
    Session = 5,
    Profile = 6,
    Replay = 7,
    TransactionProcessed = 8,
    TransactionIndexed = 9,
    Monitor = 10,
    ProfileIndexed = 11,
    Span = 12,
    MonitorSeat = 13,
    UserReportV2 = 14,
    MetricBucket = 15,
    SpanIndexed = 16,
    ProfileDuration = 17,
    ProfileChunk = 18,
    MetricSecond = 19,
    DoNotUseReplayVideo = 20,
    Uptime = 21,
    AttachmentItem = 22,
    LogItem = 23,
    LogByte = 24,
    ProfileDurationUi = 25,
    ProfileChunkUi = 26,
    SeerAutofix = 27,
    SeerScanner = 28,
    PreventUser = 29,
    PreventReview = 30,
    SizeAnalysis = 31,
    InstallableBuild = 32,
    TraceMetric = 33,
    SeerUser = 34,
    Unknown = -1,

}

Classifies the type of data that is being ingested.

Variants

Default = 0

Reserved and unused.

SDK rate limiting behavior: ignore.

Error = 1

Error events and Events with an event_type not explicitly listed below.

SDK rate limiting behavior: apply to the entire envelope if it contains an item type event.

Transaction = 2

Transaction events.

SDK rate limiting behavior: apply to the entire envelope if it contains an item transaction.

Security = 3

Events with an event type of csp, hpkp, expectct and expectstaple.

SDK rate limiting behavior: ignore.

Attachment = 4

An attachment. Quantity is the size of the attachment in bytes.

SDK rate limiting behavior: apply to all attachments.

Session = 5

Session updates. Quantity is the number of updates in the batch.

SDK rate limiting behavior: apply to all sessions and session aggregates.

Profile = 6

Profile

This is the category for processed profiles (all profiles, whether or not we store them).

SDK rate limiting behavior: apply to all profiles.

Replay = 7

Session Replays

SDK rate limiting behavior: apply to all Session Replay data.

TransactionProcessed = 8

DEPRECATED: A transaction for which metrics were extracted.

This category is now obsolete because the Transaction variant will represent processed transactions from now on.

SDK rate limiting behavior: ignore.

TransactionIndexed = 9

Indexed transaction events.

This is the category for transaction payloads that were accepted and stored in full. In contrast, transaction only guarantees that metrics have been accepted for the transaction.

SDK rate limiting behavior: ignore.

Monitor = 10

Monitor check-ins.

SDK rate limiting behavior: apply to items of type check_in.

ProfileIndexed = 11

Indexed Profile

This is the category for indexed profiles that will be stored later.

SDK rate limiting behavior: ignore.

Span = 12

Span

This is the category for spans from which we extracted metrics from.

SDK rate limiting behavior: apply to spans that are not sent in a transaction.

MonitorSeat = 13

Monitor Seat

Represents a monitor job that has scheduled monitor checkins. The seats are not ingested but we define it here to prevent clashing values since this data category enumeration is also used outside of Relay via the Python package.

SDK rate limiting behavior: ignore.

UserReportV2 = 14

User Feedback

Represents a User Feedback processed. Currently standardized on name UserReportV2 to avoid clashing with the old UserReport. TODO(jferg): Rename this to UserFeedback once old UserReport is deprecated.

SDK rate limiting behavior: apply to items of type ‘feedback’.

MetricBucket = 15

Metric buckets.

SDK rate limiting behavior: apply to statsd and metrics items.

SpanIndexed = 16

SpanIndexed

This is the category for spans we store in full.

SDK rate limiting behavior: ignore.

ProfileDuration = 17

ProfileDuration

This data category is used to count the number of milliseconds per indexed profile chunk, excluding UI profile chunks.

SDK rate limiting behavior: apply to profile chunks.

ProfileChunk = 18

ProfileChunk

This is a count of profile chunks received. It will not be used for billing but will be useful for customers to track what’s being dropped.

SDK rate limiting behavior: apply to profile chunks.

MetricSecond = 19

MetricSecond

Reserved by billing to summarize the bucketed product of metric volume and metric cardinality. Defined here so as not to clash with future categories.

SDK rate limiting behavior: ignore.

DoNotUseReplayVideo = 20

Replay Video

This is the data category for Session Replays produced via a video recording.

SDK rate limiting behavior: ignore.

Uptime = 21

This is the data category for Uptime monitors.

SDK rate limiting behavior: ignore.

AttachmentItem = 22

Counts the number of individual attachments, as opposed to the number of bytes in an attachment.

SDK rate limiting behavior: apply to attachments.

LogItem = 23

LogItem

This is the category for logs for which we store the count log events for users for measuring missing breadcrumbs, and count of logs for rate limiting purposes.

SDK rate limiting behavior: apply to logs.

LogByte = 24

LogByte

This is the category for logs for which we store log event total bytes for users.

SDK rate limiting behavior: apply to logs.

ProfileDurationUi = 25

Profile duration of a UI profile.

This data category is used to count the number of milliseconds per indexed UI profile chunk.

See also: Self::ProfileDuration

SDK rate limiting behavior: apply to profile chunks.

ProfileChunkUi = 26

UI Profile Chunk.

This data category is used to count the number of milliseconds per indexed UI profile chunk.

See also: Self::ProfileChunk

SDK rate limiting behavior: apply to profile chunks.

SeerAutofix = 27

This is the data category to count Seer Autofix run events.

SDK rate limiting behavior: ignore.

SeerScanner = 28

This is the data category to count Seer Scanner run events.

SDK rate limiting behavior: ignore.

PreventUser = 29

DEPRECATED: Use SeerUser instead.

PreventUser

This is the data category to count the number of assigned Prevent Users.

SDK rate limiting behavior: ignore.

PreventReview = 30

PreventReview

This is the data category to count the number of Prevent review events.

SDK rate limiting behavior: ignore.

SizeAnalysis = 31

Size analysis

This is the data category to count the number of size analyses performed. ‘Size analysis’ a static binary analysis of a preprod build artifact (e.g. the .apk of an Android app or MacOS .app). When enabled there will typically be one such analysis per uploaded artifact.

SDK rate limiting behavior: ignore.

InstallableBuild = 32

InstallableBuild

This is the data category to count the number of installable builds. It counts the number of artifacts uploaded not the number of times the artifacts are downloaded for installation. When enabled there will typically be one ‘InstallableBuild’ per uploaded artifact.

SDK rate limiting behavior: ignore.

TraceMetric = 33

TraceMetric

This is the data category to count the number of trace metric items.

SeerUser = 34

SeerUser

This is the data category to count the number of Seer users.

SDK rate limiting behavior: ignore.

Unknown = -1

Any other data category not known by this Relay.

Implementations

impl DataCategory

pub fn from_name(string: &str) -> DataCategory

Returns the data category corresponding to the given name.

pub fn name(self) -> &'static str

Returns the canonical name of this data category.

pub fn is_error(self) -> bool

Returns true if the DataCategory refers to an error (i.e an error event).

pub fn value(self) -> Option<u8>

Returns the numeric value for this outcome.

pub fn index_category(self) -> Option<DataCategory>

Returns a dedicated category for indexing if this data can be converted to metrics.

This returns None for most data categories.

pub fn is_indexed(self) -> bool

Returns true if this data category is an indexed data category.