Struct relay_quotas::Quota

pub struct Quota {
    pub id: Option<String>,
    pub categories: DataCategories,
    pub scope: QuotaScope,
    pub scope_id: Option<String>,
    pub limit: Option<u32>,
    pub window: Option<u64>,
    pub reason_code: Option<ReasonCode>,
}

Configuration for a data ingestion quota (rate limiting).

Sentry applies multiple quotas to incoming data before accepting it, some of which can be configured by the customer. Each piece of data (such as event, attachment) will be counted against all quotas that it matches with based on the category.

Fields
id: Option<String>

The unique identifier for counting this quota. Required, except for quotas with a limit of 0, since they are statically enforced.

categories: DataCategories

A set of data categories that this quota applies to. If missing or empty, this quota applies to all data.

scope: QuotaScope

A scope for this quota. This quota is enforced separately within each instance of this scope (e.g. for each project key separately). Defaults to QuotaScope::Organization.

scope_id: Option<String>

Identifier of the scope to apply to. If set, then this quota will only apply to the specified scope instance (e.g. a project key). Requires scope to be set explicitly.

limit: Option<u32>

Maximum number of matching events allowed. Can be 0 to reject all events, None for an unlimited counted quota, or a positive number for enforcement. Requires window if the limit is not 0.

window: Option<u64>

The time window in seconds to enforce this quota in. Required in all cases except limit=0, since those quotas are not measured.

reason_code: Option<ReasonCode>

A machine readable reason returned when this quota is exceeded. Required in all cases except limit=None, since unlimited quotas can never be exceeded.

Implementations

impl Quota

pub fn is_valid(&self) -> bool

Returns whether this quota is valid for tracking.

There are a few conditions at which quotas are invalid:

The quota only applies to Unknown data categories.

The quota is counted (not limit 0) but specifies categories with different units.

pub fn matches(&self, scoping: ItemScoping<'_>) -> bool

Checks whether the quota’s constraints match the current item.