relay_quotas
Struct RateLimit

pub struct RateLimit {
    pub categories: DataCategories,
    pub scope: RateLimitScope,
    pub reason_code: Option<ReasonCode>,
    pub retry_after: RetryAfter,
    pub namespaces: SmallVec<[MetricNamespace; 1]>,
}

A bounded rate limit.

Fields
categories: DataCategories

A set of data categories that this quota applies to. If missing or empty, this rate limit applies to all data.

scope: RateLimitScope

The scope of this rate limit.

reason_code: Option<ReasonCode>

A machine readable reason indicating which quota caused it.

retry_after: RetryAfter

A marker when this rate limit expires.

namespaces: SmallVec<[MetricNamespace; 1]>

The metric namespace of this rate limit.

Ignored on all data categories except for MetricBucket. If empty, this rate limit applies to metrics of all namespaces.

Implementations

impl RateLimit

pub fn from_quota( quota: &Quota, scoping: &Scoping, retry_after: RetryAfter, ) -> Self

Creates a new rate limit for the given Quota.

pub fn matches(&self, scoping: ItemScoping<'_>) -> bool

Checks whether the rate limit applies to the given item.