Crate relay_quotas

Quotas and rate limiting for Relay.

Structs

CachedRateLimits

Like RateLimits, a collection of scoped rate limits but with all the checks necessary to cache the limits.

ItemScoping

Data categorization and scoping information.

Quota

Configuration for a data ingestion quota (rate limiting).

RateLimit

A bounded rate limit.

RateLimits

A collection of scoped rate limits.

RateLimitsIntoIter

An iterator that moves out of RateLimtis.

RateLimitsIter

Immutable rate limits iterator.

ReasonCode

A machine readable, freeform reason code for rate limits.

RedisRateLimiter

A service that executes quotas and checks for rate limits in a shared cache.

RetryAfter

A monotonic expiration marker for RateLimits.

Scoping

Data scoping information.

Enums

DataCategory

Classifies the type of data that is being ingested.

InvalidRetryAfter

Error parsing a RetryAfter.

MetricNamespaceScoping

Item scoping of metric namespaces.

QuotaScope

The scope that a quota applies to.

RateLimitScope

The scope that a rate limit applied to.

RateLimitingError

An error returned by RedisRateLimiter.

Type Aliases

DataCategories

An efficient container for data categories that avoids allocations.